feat: rewrite in typescript
With this commit I've rewritten the micro-post module using typescript.
This change has the benefit of providing maintainable typescript type definitions which evolve as the module itself does and are automatically installed when the module gets installed.

The change does not require any modification by the module users and functions as a drop in replacement, the tests are the same (except type annotations) and still pass.