feat: let micro handle the response encoding
micro is able to encode various data formats like: text, json, buffer and streams

Delegating the encoding of the response to micro opens up the use-cases where micro-post is applicable and also reduced redundancy of encoding implementations.
This branch includes an extension with which it is possible to pass a default micro RequestHandler as response handler function, profiting too of micros encoding capabilities.
Using the test suite I confirmed these changes are backwards compatible.